scripts/requirements.txt
psycopg2-binary
SQLAlchemy
python-dotenv
python-jose
passlib
ankipandas
fsrs
